Intermediate and failing version of making nom work
However, it makes it way more complicated than it has to be.
Also nom will not return any result in case the parsing is incomplete,
which may not be what we actually want! Partial indexing is fine,
in case of multiple rows for example, and right now we are forced
to continue parsing from the beginning at all times.

Therefore parsing should just be implemented so that it will only
index the byte stream without doing any unnecessary decoding work.