Decouple direction from protocol version
That way, it will be easier to provide the version
to the encoder.

Overall plan is to have one datatype to support all
protocols.
The bad thing about this is that the type system will
not prevent errors.

So maybe it will be done using a type parameter, or hand-crafted
types?