Remove type-based version differentiation
The idea is that with a type-based implementation,
code-reuse will be more boilerplaty.

Also I think it will be easier to implement behaviours
like 'prefer v4, but allow v3' if it's more of a runtime thing.

Thanks to the ProtocolVersion enum, it's quite easy to handle
special cases, and only special cases.