Allow both client implementation in parallel
Reason is that using the streaming client is more difficult.
Thus it's best to just support both implementation, which fortunately
will work with similar datatypes 'under the hood'.

However, clients with the need to stream will always have to
differentiate results with either a complete reply, or a stream
of chunks.