Service trait for streaming client handle
The idea is to provide a simple interface to the user
of the API - for now, we simply don't support streaming
requests, but do so with streaming responses.

At some later point that can easily be changed, for example
when we know how queries work (imagine those that populate a
database with gigabytes of data). The latter would be
problematic due to the maxium frame size of 256MB, which is
arbitrarily imposed by cassandra it seems, at least for responses.