And here comes the Client :)!!
The good news: all the pieces seem to fit together now.
Bad new: where is the test-drivenness?
Hints for how to test this are in tokio-proto, but for now
I just rely on integration tests, to be done by the CLI,
with a real server.
Given the strong type system we have on our side, that should
actually be quite fine!