Box client-proxy to support various implementations
NOTE: I also removed all the .boxed() invocation and replaced
them with Box::new(), as the latter does not bring in the Send
requirement, which really isn't needed as Threads are not implied
by any of this.
Boxed() should also not be used I guess, as least not in libraries.